JSGARVEY US220 Separate Classes for clarification
